merge: pkg lock
