+++ src/hooks/airdrop/stateHelpers/useAirdropTokensRefresh.ts
@@ -1,7 +1,6 @@
 import { AirdropTokens, setAirdropTokens, useAirdropStore } from '@app/store/useAirdropStore';
 import { useEffect } from 'react';
-import { getCurrentWindow } from '@tauri-apps/api/window';
-const currentWindow = getCurrentWindow();
+
 export async function fetchAirdropTokens(airdropApiUrl: string, airdropTokens: AirdropTokens) {
     const response = await fetch(`${airdropApiUrl}/auth/local/refresh`, {
         method: 'POST',
@@ -12,7 +11,6 @@ export async function fetchAirdropTokens(airdropApiUrl: string, airdropTokens: A
             refreshToken: airdropTokens.refreshToken,
         }),
     });
-    console.debug(response);
     if (!response.ok) {
         throw new Error('Failed to refresh token');
     }
@@ -21,34 +19,22 @@ export async function fetchAirdropTokens(airdropApiUrl: string, airdropTokens: A
     return data;
 }
 
-export async function handleRefreshAirdropTokens(airdropApiUrl: string, emit = false) {
+export async function handleRefreshAirdropTokens(airdropApiUrl: string) {
     const airdropTokens = useAirdropStore.getState().airdropTokens;
-    const syncedAidropWithBackend = useAirdropStore.getState().syncedWithBackend;
     let fetchedAirdropTokens: AirdropTokens | undefined;
-    let startupToken: AirdropTokens['token'] | undefined;
     // 5 hours from now
     const expirationLimit = new Date(new Date().getTime() + 1000 * 60 * 60 * 5);
     const tokenExpirationTime = airdropTokens?.expiresAt && new Date(airdropTokens?.expiresAt * 1000);
 
     const tokenHasExpired = tokenExpirationTime && tokenExpirationTime < expirationLimit;
-    console.debug(airdropTokens);
-    console.debug(`tokenHasExpired= ${tokenHasExpired}`);
-    console.debug(`syncedAidropWithBackend= ${syncedAidropWithBackend}`);
-    if (airdropTokens && (!syncedAidropWithBackend || tokenHasExpired)) {
+    if (airdropTokens && tokenHasExpired) {
         try {
             fetchedAirdropTokens = await fetchAirdropTokens(airdropApiUrl, airdropTokens);
-            if (fetchedAirdropTokens?.token) {
-                startupToken = fetchedAirdropTokens.token;
-            }
         } catch (error) {
             console.error('Error refreshing airdrop tokens:', error);
         }
     }
 
-    if (emit && startupToken) {
-        console.debug('emitting');
-        await currentWindow.emit('startup-token', startupToken);
-    }
     await setAirdropTokens(fetchedAirdropTokens);
 }
 export function useAirdropTokensRefresh() {
